pyproject.toml
[tool.poetry]
name = "coursework-6"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
black = "^24.4.2"
ipython = "^8.26.0"
django-apscheduler = "^0.6.2"
pillow = "^10.4.0"
pytils = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
